Add decode_xml_wineventlog processor
