+++ libbeat/docs/processors-list.asciidoc
@@ -63,7 +63,10 @@ ifndef::no_decode_json_fields_processor[]
 * <<decode-json-fields,`decode_json_fields`>>
 endif::[]
 ifndef::no_decode_xml_processor[]
-* <<decode_xml, `decode_xml`>>
+* <<decode-xml, `decode_xml`>>
+endif::[]
+ifndef::no_decode_xml_wineventlog_processor[]
+* <<decode-xml-wineventlog, `decode_xml_wineventlog`>>
 endif::[]
 ifndef::no_decompress_gzip_field_processor[]
 * <<decompress-gzip-field,`decompress_gzip_field`>>
@@ -183,6 +186,12 @@ endif::[]
 ifndef::no_decode_json_fields_processor[]
 include::{libbeat-processors-dir}/actions/docs/decode_json_fields.asciidoc[]
 endif::[]
+ifndef::no_decode_xml_processor[]
+include::{libbeat-processors-dir}/decode_xml/docs/decode_xml.asciidoc[]
+endif::[]
+ifndef::no_decode_xml_wineventlog_processor[]
+include::{libbeat-processors-dir}/decode_xml_wineventlog/docs/decode_xml_wineventlog.asciidoc[]
+endif::[]
 ifndef::no_decompress_gzip_field_processor[]
 include::{libbeat-processors-dir}/actions/docs/decompress_gzip_field.asciidoc[]
 endif::[]
@@ -233,9 +242,6 @@ include::{libbeat-processors-dir}/actions/docs/truncate_fields.asciidoc[]
 endif::[]
 ifndef::no_urldecode_processor[]
 include::{libbeat-processors-dir}/urldecode/docs/urldecode.asciidoc[]
-endif::[]
-ifndef::no_decode_xml_processor[]
-include::{libbeat-processors-dir}/decode_xml/docs/decode_xml.asciidoc[]
 endif::[]
 
 //# end::processors-include[]
