--- libbeat/processors/decode_xml/schema.go
@@ -1,87 +1,0 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package decode_xml
-
-import (
-	"bytes"
-	"errors"
-
-	"github.com/elastic/beats/v7/libbeat/common"
-	"github.com/elastic/beats/v7/libbeat/common/encoding/xml"
-	"github.com/elastic/beats/v7/libbeat/logp"
-)
-
-const wineventlogSchema = "wineventlog"
-
-type newDecoderFunc func(cfg decodeXMLConfig) decoder
-
-// a decoder will decode the raw XML into a set of fields,
-// optionally will return additional ECS mappings when possible
-type decoder func(p []byte) (decodedXML, ecsFields common.MapStr, err error)
-
-var (
-	registeredDecoders                = map[string]newDecoderFunc{}
-	newDefaultDecoder  newDecoderFunc = newSchemaLessDecoder
-)
-
-func registerDecoder(schema string, dec newDecoderFunc) error {
-	if schema == "" {
-		return errors.New("schema can't be empty")
-	}
-
-	if dec == nil {
-		return errors.New("decoder can't be nil")
-	}
-
-	if _, found := registeredDecoders[schema]; found {
-		return errors.New("already registered")
-	}
-
-	registeredDecoders[schema] = dec
-
-	return nil
-}
-
-func newDecoder(cfg decodeXMLConfig) decoder {
-	newDec, found := registeredDecoders[cfg.Schema]
-	if !found {
-		return newDefaultDecoder(cfg)
-	}
-	return newDec(cfg)
-}
-
-func registerDecoders() {
-	log := logp.L().Named(logName)
-	log.Debug(registerDecoder(wineventlogSchema, newWineventlogDecoder))
-}
-
-func newSchemaLessDecoder(cfg decodeXMLConfig) decoder {
-	return func(p []byte) (common.MapStr, common.MapStr, error) {
-		dec := xml.NewDecoder(bytes.NewReader(p))
-		if cfg.ToLower {
-			dec.LowercaseKeys()
-		}
-
-		out, err := dec.Decode()
-		if err != nil {
-			return nil, nil, err
-		}
-
-		return common.MapStr(out), nil, nil
-	}
-}
